Refactor to function interfaces for int and float

+++ src/main/java/ar/com/kfgodel/function/intertype/floats/FloatToBooleanFunction.java
@@ -3,16 +3,15 @@ package ar.com.kfgodel.function.intertype.floats;
 import java.util.function.Function;
 
 /**
- * This type defines an operator for primitive boolean types
  * Created by kfgodel on 23/07/17.
  */
-public interface FloatToBooleanFunction extends Function<Float, Boolean> {
+public interface FloatToBooleanFunction extends Function<Integer, Boolean> {
 
   boolean apply(float aBoolean);
 
   @Override
-  default Boolean apply(Float aBoolean){
-    return this.apply(aBoolean.floatValue());
+  default Boolean apply(Integer aBoolean){
+    return this.apply(aBoolean.intValue());
   }
 
 }
