+++ src/main/java/ar/com/kfgodel/function/intertype/floats/FloatToDoubleFunction.java
@@ -3,15 +3,14 @@ package ar.com.kfgodel.function.intertype.floats;
 import java.util.function.Function;
 
 /**
- * This type represents a function that takes a boolean input and produces a double as result
  * Date: 29/07/17 - 19:57
  */
-public interface FloatToDoubleFunction extends Function<Float, Double> {
+public interface FloatToDoubleFunction extends Function<Integer, Double> {
 
   double apply(float input);
 
   @Override
-  default Double apply(Float input){
-    return apply(input.floatValue());
+  default Double apply(Integer input){
+    return apply(input.intValue());
   }
 }
